backend of difficulty rating of subject is completed, not loading the css inside the content in reivew page is solved

+++ app/database/migrations/2015_01_27_012541_create_subject_table.php
@@ -16,7 +16,7 @@ class CreateSubjectTable extends Migration {
 		{
 			$table->increments('id');
             $table->string('subject_name',100);    
-            $table->integer('rate')->default(0);
+            $table->float('rate')->default(0);
             $table->integer('number')->default(0);  
             $table->timestamps();
 		});
